+++ TR Poser/Parsing/TR2 Level Elements/TR2FrameData.m
@@ -42,9 +42,9 @@
 	frame.boundingBoxMaxY = self.frameData[position++];
 	frame.boundingBoxMaxZ = self.frameData[position++];
 	
-	frame.offsetX = self.frameData[position++];
-	frame.offsetY = self.frameData[position++];
-	frame.offsetZ = self.frameData[position++];
+	frame.offsetX = (int16_t) self.frameData[position++];
+	frame.offsetY = (int16_t) self.frameData[position++];
+	frame.offsetZ = (int16_t) self.frameData[position++];
 	
 	for (NSUInteger i = 0; i < object.meshCount; i++)
 	{
